Moving the base DataRun class to its own file.
